requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
filelock==3.24.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grandalf==0.8
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-google-genai==4.2.1
langchain-huggingface==1.2.0
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
openai==2.23.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyparsing==3.3.2
pypdf==6.8.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
